refactor(policy): establish policy/ — egress/attachment, egress/result, guardrail
Next lane after channels/ + delivery/, smallest-blast-radius-first per the
owner's ordering. send_allowlist -> policy/egress/attachment (which local
paths may egress), team_result_guard -> policy/egress/result (tier-gated
result egress), team_guardrail -> policy/guardrail (agent policy, not
transport). All three keep true sys.modules aliases at their retired names.

Vendored ag2-sparrow imports keep the FLAT names by design, so the two
'from ag2_sparrow import <flat>' sites are deliberately NOT rewritten; the
vendor map's source paths follow the move and --check stays in sync.
