+++ tests/discord-bridge-allowlist.test.py
@@ -35,7 +35,7 @@ import sys
 
 REPO = Path(__file__).resolve().parent.parent
 BRIDGE = REPO / "src" / "discord-bridge.py"
-ALLOWLIST_MODULE = REPO / "src" / "send_allowlist.py"
+ALLOWLIST_MODULE = REPO / "src" / "policy" / "egress" / "attachment.py"
 
 
 def fail(msg: str, context: str = "") -> int:
@@ -63,7 +63,7 @@ def main() -> int:
     found_in = None
     for candidate_src, label in (
         (src, "discord-bridge.py"),
-        (ALLOWLIST_MODULE.read_text() if ALLOWLIST_MODULE.exists() else "", "send_allowlist.py"),
+        (ALLOWLIST_MODULE.read_text() if ALLOWLIST_MODULE.exists() else "", "policy/egress/attachment.py"),
     ):
         m = HELPER_RE.search(candidate_src)
         if m:
@@ -73,22 +73,22 @@ def main() -> int:
     if helper_body is None:
         return fail(
             "`_is_path_sendable` / `is_path_sendable` function not found in either "
-            "src/discord-bridge.py or src/send_allowlist.py"
+            "src/discord-bridge.py or src/policy/egress/attachment.py"
         )
     # If the implementation lives in send_allowlist.py, discord-bridge
     # must import it (otherwise the file-send sites have a dangling
     # name). Verify the import is present.
-    if found_in == "send_allowlist.py":
-        # The import may use `from send_allowlist import is_path_sendable`
-        # OR `from send_allowlist import (is_path_sendable as _alias, ...)`
+    if found_in == "policy/egress/attachment.py":
+        # The import may use `from policy.egress.attachment import is_path_sendable`
+        # OR `from policy.egress.attachment import (is_path_sendable as _alias, ...)`
         # — both shapes satisfy the contract. Use DOTALL so multi-line
         # parenthesized imports match.
         if not re.search(
-            r"from\s+send_allowlist\s+import[\s\S]*?is_path_sendable",
+            r"from\s+policy\.egress\.attachment\s+import[\s\S]*?is_path_sendable",
             src,
         ):
             return fail(
-                "send_allowlist.py defines is_path_sendable but discord-bridge.py "
+                "policy/egress/attachment.py defines is_path_sendable but discord-bridge.py "
                 "does not import it — the gate sites would reference an undefined name."
             )
